Split up loadCommands
- Split up command functions to prepare for message functions.
