+++ src/util.js
@@ -3,5 +3,6 @@ module.exports = {};
 //
 
 // - loading modules
-module.exports.load = require("./loaddata.js")
-module.exports.checkUpdate = require("./checkUpdate.js")
+module.exports.load = require("./loaddata.js");
+module.exports.checkUpdate = require("./checkUpdate.js");
+module.exports.loadCommands = require("./loadCommands.js");
